Add assembly files and namespace to SoundManager and IAudioControl
