+++ Assets/_Dev/Scripts/Manager/Sounds.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5fac7e9269cb040e693945219a4af78e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
