+++ Assets/_Dev/Scripts/Manager/Sounds/WheelMaker.Manager.Sounds.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0c8de95abebc54f7786dca948e9acab4
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
