+++ Assets/_Dev/Scripts/Scene/Title/WheelMaker.Scene.Title.asmdef
@@ -1,0 +1,17 @@
+{
+    "name": "WheelMaker.Scene.Title",
+    "rootNamespace": "",
+    "references": [
+        "WheelMaker.Manager.Interfaces",
+        "WheelMaker.Manager.Sounds"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
